requirements.txt
apturl==0.5.2
asttokens==2.0.8
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
entrypoints==0.3
executing==0.10.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
gpg===1.13.1-unknown
greenlet==0.4.15
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.12.0
incremental==16.10.1
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lightdm-gtk-greeter-settings==1.2.2
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
menulibre==2.2.1
more-itertools==4.2.0
msgpack==0.6.2
mugshot==0.4.2
neovim==0.3.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
packaging==20.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.5.1
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rofication==1.2.2
SecretStorage==2.3.1
service-identity==18.1.0
sgt-launcher==0.2.5
simplejson==3.16.0
six==1.14.0
sos==4.3
SQLAlchemy==1.4.40
ssh-import-id==5.10
stack-data==0.4.0
systemd-python==234
traitlets==5.3.0
Twisted==18.9.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.2.2
xcffib==0.8.1
xkit==0.0.0
youtube-dl==2021.6.6
zipp==1.0.0
zope.interface==4.7.1
